fix: 更新 passless-server
1. 移除廢棄檔案
2. 將 router 內部 function 移動至 controllers 內
