+++ apps/server/src/controllers/authentication/index.ts
@@ -1,0 +1,1 @@
+export * from "./authentication";
