--- apps/server/src/controllers/helpers/decodeAttestationObject.ts
@@ -1,40 +1,0 @@
-import { Base64Url } from '../../utils';
-import { isoCBOR } from '@simplewebauthn/server/helpers';
-
-export type AttestationFormat =
-  | 'fido-u2f'
-  | 'packed'
-  | 'android-safetynet'
-  | 'android-key'
-  | 'tpm'
-  | 'apple'
-  | 'none';
-
-// instance of `Map`
-export type AttestationObject = {
-  get(key: 'fmt'): AttestationFormat;
-  get(key: 'attStmt'): AttestationStatement;
-  get(key: 'authData'): Uint8Array;
-};
-
-/**
- * `AttestationStatement` will be an instance of `Map`, but these keys help make finite the list of
- * possible values within it.
- */
-export type AttestationStatement = {
-  get(key: 'sig'): Uint8Array | undefined;
-  get(key: 'x5c'): Uint8Array[] | undefined;
-  get(key: 'response'): Uint8Array | undefined;
-  get(key: 'alg'): number | undefined;
-  get(key: 'ver'): string | undefined;
-  get(key: 'certInfo'): Uint8Array | undefined;
-  get(key: 'pubArea'): Uint8Array | undefined;
-  // `Map` properties
-  readonly size: number;
-};
-
-export function decodeAttestationObject(attestationObject: string) {
-  return isoCBOR.decodeFirst<AttestationObject>(
-    new Uint8Array(Base64Url.decodeBase64Url(attestationObject))
-  );
-}
