+++ apps/server/src/controllers/index.ts
@@ -1,0 +1,3 @@
+export * from "./registration";
+export * from "./authentication";
+export * from "./passkeys";
