+++ apps/server/src/controllers/passkeys/index.ts
@@ -1,0 +1,1 @@
+export * from "./passkey";
