+++ apps/server/src/controllers/passkeys/passkey.ts
@@ -1,0 +1,90 @@
+import express, { Request, Response, NextFunction } from "express";
+import {
+  generateAuthenticationOptions,
+  verifyAuthenticationResponse
+} from "@simplewebauthn/server";
+import type { GenerateAuthenticationOptionsOpts } from "@simplewebauthn/server";
+import type { AuthenticationResponseJSON } from "@webauthn/types";
+
+import { TypedRequestBody } from "../../types";
+import { CustomError } from "../../middleware";
+import { credentialService } from "../../service";
+import { Base64Url } from "../../utils";
+
+export const handlePasskeysStart = async (req: Request, res: Response, next: NextFunction) => {
+  try {
+    const opts: GenerateAuthenticationOptionsOpts = {
+      userVerification: "preferred",
+      allowCredentials: [],
+      rpID: process.env.RP_ID
+    };
+
+    const options = await generateAuthenticationOptions(opts);
+
+    req.session.currentChallenge = options.challenge;
+
+    res.status(200).json({
+      status: "Success",
+      data: options
+    });
+  } catch (error) {
+    next(error instanceof CustomError ? error : new CustomError("Internal Server Error", 500));
+  }
+};
+
+type PutPasskeyReqBody = TypedRequestBody<{
+  data: AuthenticationResponseJSON;
+}>;
+
+export const handlePasskeysFinish = async (
+  req: PutPasskeyReqBody,
+  res: Response,
+  next: NextFunction
+) => {
+  const { currentChallenge } = req.session;
+
+  if (!currentChallenge) {
+    return next(new CustomError("Current challenge is missing", 400));
+  }
+
+  try {
+    const { data = undefined } = req.body;
+
+    if (!data) {
+      return next(new CustomError("缺少必要資訊", 403));
+    }
+
+    const authenticator = await credentialService.getCredentialByCredentialId(data.id);
+    if (!authenticator) {
+      return next(new CustomError("User is not registered this device", 403));
+    }
+
+    const verification = await verifyAuthenticationResponse({
+      response: data,
+      expectedChallenge: currentChallenge,
+      expectedOrigin: String(req.headers.origin),
+      expectedRPID: process.env.RP_ID,
+      authenticator: {
+        credentialID: new Uint8Array(Base64Url.decodeBase64Url(authenticator.credential_id)),
+        credentialPublicKey: new Uint8Array(Base64Url.decodeBase64Url(authenticator.public_key)),
+        counter: authenticator.counter,
+        transports: JSON.parse(authenticator.transports)
+      },
+      requireUserVerification: true
+    });
+
+    if (verification.verified && verification.authenticationInfo) {
+    } else {
+      next(new CustomError("Verification failed", 400));
+    }
+
+    res.status(200).json({
+      status: "Success",
+      data: {}
+    });
+  } catch (error) {
+    next(error instanceof CustomError ? error : new CustomError("Internal Server Error", 500));
+  } finally {
+    req.session.currentChallenge = undefined;
+  }
+};
