--- apps/server/src/controllers/verification/index.ts
@@ -1,2 +1,0 @@
-export * from './verifyRegistrationResponse';
-export * from '../helpers/decodeClientDataJSON';
