+++ apps/server/src/middleware/errorHandler.ts
@@ -1,12 +1,8 @@
-import { Request, Response, NextFunction } from "express";
+import { Request, Response } from "express";
 import { CustomError } from "./customError";
 
-interface ErrorWithStatus extends Error {
-  statusCode?: number;
-}
-
-export function handleError(err: CustomError, req: Request, res: Response, next: NextFunction) {
+export function handleError(err: CustomError, req: Request, res: Response) {
   const statusCode = err.statusCode || 500;
   const message = err.message || "Internal Server Error";
-  res.status(statusCode).send({ error: message });
+  res.status(statusCode).json({ status: "Error", message: message });
 }
