+++ packages/@webauthb-types/src/types/ResponseJSON/AuthenticationResponseJSON.type.ts
@@ -1,11 +1,12 @@
+import { Base64UrlString } from "../common";
 import type { AuthenticationExtensionsClientOutputsJSON } from "./common.type";
 
 /**
  * 前端產生 signature 後轉換為 JSON 格式
  */
 export interface AuthenticationResponseJSON {
-  id: string;
-  rawId: string;
+  id: Base64UrlString;
+  rawId: Base64UrlString;
   response: AuthenticatorAssertionResponseJSON;
   authenticatorAttachment?: AuthenticatorAttachment;
   clientExtensionResults: AuthenticationExtensionsClientOutputsJSON;
@@ -20,8 +21,8 @@ type OmitAuthenticatorAssertionResponse =
 
 export interface AuthenticatorAssertionResponseJSON
   extends Omit<AuthenticatorAssertionResponse, OmitAuthenticatorAssertionResponse> {
-  clientDataJSON: string;
-  authenticatorData: string;
-  signature: string;
-  userHandle?: string;
+  clientDataJSON: Base64UrlString;
+  authenticatorData: Base64UrlString;
+  signature: Base64UrlString;
+  userHandle?: Base64UrlString;
 }
