+++ packages/@webauthb-types/src/types/common/base64url.ts
@@ -1,0 +1,1 @@
+export type Base64UrlString = string;
