+++ packages/@webauthb-types/src/types/index.ts
@@ -1,2 +1,3 @@
+export type * from "./common";
 export type * from "./CredentialOptionsJSON";
 export type * from "./ResponseJSON";
